% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsTMT.gui.R
\docType{package}
\name{MSstatsTMT.gui}
\alias{MSstatsTMT.gui}
\title{MSstatsTMT.gui Package}
\description{
Contains a shiny app for MSstatsTMT r package, in a user friendly way.
Also contains three other functions, including 2 based on the dataProcessPlotsTMT function from the
package MSstatsTMT
}
\author{
Marc-Antoine GERAULT \email{marco.gerault@gmail.com}
}
