% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plo_volc.R
\name{plo_volc}
\alias{plo_volc}
\title{Volcano plot}
\usage{
plo_volc(
  data,
  lim_pv = 0.05,
  lim_dif = c(-1, 1),
  correction = FALSE,
  perf_corr = FALSE,
  your_corr = "none",
  comp = "0.125-Norm",
  curve = FALSE,
  curvature = 0.1,
  tit = "mqpar25 MSstatsTMT",
  ytit = "-log10(p-value)"
)
}
\arguments{
\item{data}{The output from \code{\link{groupComparionTMT}}}

\item{lim_pv}{The p-value threshold}

\item{lim_dif}{A double, corresponding to the two log2 fold-change thresholds}

\item{correction}{A logical, if TRUE will use the p-v.adjust column from the output from \code{\link{groupComparionTMT}}}

\item{perf_corr}{A logical, if TRUE will correct the p-values using the method from p.adjust}

\item{your_corr}{A character vector, if perf_corr TRUE will use this method for correct the p-values;
it will print the mehtod used for correcting the p-values on the plot}

\item{comp}{A character vector, indicating the comparison you want to see the volcano plot. Has to match one of the element of
'Label' column from the data}

\item{curve}{A logical, if TRUE will plot a curvature on the volcano plot}

\item{curvature}{A numeric, indicating the value of the curvature}

\item{tit}{The title of the volcano plot}

\item{ytit}{The title of the y axis}
}
\value{
A volcano plot
}
\description{
\code{plo_volc} Plot a volcano from the output data from the function \code{\link{groupComparionTMT}}
}
\examples{
library(MSstatsTMT)
quant.pd.msstats <- proteinSummarization(input.pd,
                                         method="msstats",
                                         global_norm=TRUE,
                                         reference_norm=TRUE)

test.pairwise <- groupComparisonTMT(quant.pd.msstats, moderated = TRUE)

plo_volc(test.pairwise)
}
\seealso{
\code{\link{groupComparionTMT}} from MSstatsTMT package for more details
}
